Use CREDENTIALS_KEY_PREFIX
